+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+* Added `channel_spec.py` as a helper and documentation
 * Auxiliary images now have `warp_aux_to_img` property
 * Base images now have `warp_img_to_vid` property